backend-server/requirements.txt
confluent-kafka
google-cloud-dataproc
google-cloud-storage